[TE] Support TianLong SHCA IB
Add USE_SHCA for TianLong SHCA NICs with extended 17-bit LIDs. Convert
LIDs via shca_17b_types.h helpers and widen lid fields to uint32_t across
Transfer Engine and TENT RDMA paths.

Gate yalantinglibs IBV RPC (YLT_ENABLE_IBV) behind USE_SHCA=OFF so SHCA
builds leave ib_socket untouched and fall back to TCP for store/RPC.

+++ mooncake-transfer-engine/include/transfer_metadata.h
@@ -44,7 +44,7 @@ class TransferMetadata {
    public:
     struct DeviceDesc {
         std::string name;
-        uint16_t lid;
+        uint32_t lid;
         std::string gid;
         std::string eid;  // for ub
     };
